# apps/backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  application.CreateProjectInput:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      tenant_id:
        type: string
    required:
    - name
    type: object
  application.UpdateProjectInput:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
    type: object
  http.AggBucketResponse:
    properties:
      count:
        type: integer
      key:
        type: string
      timestamp:
        type: string
    type: object
  http.AggregateRequest:
    properties:
      from:
        type: string
      group_by:
        type: string
      interval:
        description: '"1m", "5m", "1h", "1d"'
        type: string
      project_id:
        type: string
      to:
        type: string
    required:
    - from
    - project_id
    - to
    type: object
  http.AggregateResponse:
    properties:
      buckets:
        items:
          $ref: '#/definitions/http.AggBucketResponse'
        type: array
    type: object
...
  http.ExportRequest:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      body_contains:
        type: string
      format:
        description: '"json" | "csv" | "ndjson"'
        type: string
      from:
        type: string
      services:
        items:
          type: string
        type: array
      severities:
        items:
          type: string
        type: array
      tenant_id:
        type: string
      to:
        type: string
    required:
    - from
    - to
    type: object
  http.ExportResponse:
    properties:
      created_at:
        type: string
      export_id:
        type: string
      status:
        type: string
    type: object
  http.ExportStatusResponse:
    properties:
      created_at:
        type: string
      download_url:
        type: string
      export_id:
        type: string
      status:
        type: string
    type: object
...
  http.LogResponse:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      body:
        type: string
      environment:
        type: string
      host:
        type: string
      ingestion_time:
        type: string
      log_id:
        type: string
      namespace:
        type: string
      project_id:
        type: string
      request_id:
        type: string
      service:
        type: string
      severity:
        type: string
      source:
        type: string
      span_id:
        type: string
      tags:
        additionalProperties:
          type: string
        type: object
      tenant_id:
        type: string
      timestamp:
        type: string
      trace_id:
        type: string
      user_id:
        type: string
    type: object
...
  http.SearchRequest:
    properties:
      cursor:
        type: string
      limit:
        type: integer
      project_id:
        type: string
      query:
        type: string
      sort:
        $ref: '#/definitions/http.Sort'
      time_range:
        $ref: '#/definitions/http.TimeRange'
    required:
    - project_id
    type: object
  http.SearchResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/http.LogResponse'
        type: array
      next_cursor:
        type: string
      took_ms:
        type: integer
      total:
        type: integer
    type: object
  http.Sort:
    properties:
      field:
        type: string
      order:
        description: '"asc" or "desc"'
        type: string
    type: object
  http.TimeRange:
    properties:
      from:
        type: string
      to:
        type: string
    type: object
...
  response.APIError:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the Logify backend API server.
  title: Logify API
  version: "1.0"
paths:
  /exports/{id}:
    get:
      description: Return the current status of an async log export job.
      parameters:
      - description: Export ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Export status
          schema:
            $ref: '#/definitions/http.ExportStatusResponse'
      security:
      - BearerAuth: []
      summary: Get export status
      tags:
      - logs
...
  /v1/logs/{id}:
    get:
      description: Retrieve a log from an accessible project. Supply project_id for
        shared projects.
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: string
      - description: Project UUID (required for shared projects)
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Log entry
          schema:
            $ref: '#/definitions/http.LogResponse'
        "400":
          description: Invalid project_id
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Log not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Fetch failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get log by ID
      tags:
      - logs
  /v1/logs/aggregate:
    post:
      consumes:
      - application/json
      description: Run a grouped/time-bucketed aggregation over logs.
      parameters:
      - description: Aggregation query
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.AggregateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Aggregation result
          schema:
            $ref: '#/definitions/http.AggregateResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Aggregate failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Aggregate logs
      tags:
      - logs
  /v1/logs/export:
    post:
      consumes:
      - application/json
      description: Start an asynchronous log export job; returns the export job ID.
      parameters:
      - description: Export query
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ExportRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Export accepted
          schema:
            $ref: '#/definitions/http.ExportResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export logs
      tags:
      - logs
  /v1/logs/search:
    post:
      consumes:
      - application/json
      description: Search logs in a project visible to the authenticated user.
      parameters:
      - description: Search query
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/http.SearchResponse'
        "400":
          description: Invalid request or missing project_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Search failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search logs
      tags:
      - logs
  /v1/projects:
    get:
      description: Return owned and shared projects, including the caller’s persisted
        project role.
      produces:
      - application/json
      responses:
        "200":
          description: Projects retrieved successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project within the authenticated tenant.
      parameters:
      - description: Project payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.CreateProjectInput'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Project already exists
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /v1/projects/{id}:
    delete:
      description: Delete a project by UUID.
      parameters:
      - description: Project ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Invalid id format
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      description: Retrieve a project by UUID.
      parameters:
      - description: Project ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project retrieved successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid id format
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update fields on an existing project.
      parameters:
      - description: Project ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Project update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.UpdateProjectInput'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
...
securityDefinitions:
  APIKeyAuth:
    description: Project API key, restricted to sending logs.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Enter Bearer followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"